requirements.txt
flask
flask_sqlalchemy
python-dotenv
pytest